Replace bootloader with self-compiled optiboot
This bootloader was compiled from optiboot git, version 814e2a8 (Fix
some typos, improve some comments). It was compiled with these options:

	make AVR_FREQ=8000000L BAUD_RATE=57600 LED=B7 atmega328

The baudrate copies that of the Arduino Pro Mini, to allow using the
board (partly, not all pins are available) to upload once the bootloader
is present. Additionally, the LED is set correctly for this board.