Introduce STM32-named pin constants (like PB0)
This adds these constants, and at the same time uses them to define the
pin assignment table (to prevent the constants from diverging from the
actual table) and to define various pin constants.

These constants are defined in an enum rather than consts, since those
are considered constants usable to initialize other constants in C.

This also corrects the button pin, which was incorrect before.