Really configure USB vidpid correctly
This removes the vid.0 / pid.0 (which the IDE uses to display a name
for a serial port) for PROTO1 and sets it for PROTO2, so a PROTO2 board
is not shown as PROTO1. It also sets the actual pid used in the build
for PROTO1, which was forgotten before (this means PROTO1 shows up as
PROTO2, but that's ok).