Move avr subdirectory to top level
Tracking multiple architectures in the same branch is probably not
ideal, so better just remove the extra "avr" subdirectory. This means
a git clone of this repository should move one layer deeper.

Also remove the .json file with releases, which will be kept in a
different branch, so it can refer to multiple architectures.