Do not keep debugger enabled in sleep modes
It turns that that this causes problems when an interrupt or exception
happens when blocking on a WFI or WFE instruction. In that case, when
DBGMCU_CR_DBG_SLEEP is enabled, r0 is not stacked correctly (0 is
written to the stack, rather than the actual value). On exception
return, this incorrect value overwrites r0, potentially causing all
kinds of havoc (this was discovered because Serial.flush() would never
return because its `this` pointer was suddenly null).

Originally, this bit DBGMCU_CR_DBG_SLEEP was only enabled when SWD was
explicitly enabled by the sketch, but since commit f1d400f (Rework
atomic primitives to allow for shorter interrupt disable code segments),
it is enabled by default on startup.

This commit disables DBGMCU_CR_DBG_SLEEP again to work around this
problem. Hopefully there is a better fix for this problem, but for now
it at least allows code to work again.