Allow reconfiguring Serial
This adds a compiletime option (settable through the Arduino IDE) that
allows configuring Serial to point to Serial1 (hardware UART) rather
than the default SerialUSB.

Additionally, a third option is added that allows reconfiguring Serial
at runtime, for even more flexibility. This third option is not the
default, though, because it does not have `Serial.begin()`, so it only
works with sketches that are specifically written to us it. See the
comments in variant.h for more details on how this works.