Increase crystal drive strength
The default low drive strength was a bit too low for the used crystal,
which means the crystal could (in theory) rarely fail to start up or
stop running.  This increases the drive strength to the value
recommended by ST.

This could hopefully explain the cases where we saw boards lockup after
a while, but testing with this patch applied still showed lockups, so
that likely has another cause. Still, setting the right drivestrength
should improve things.